# cloud-layer/cloud-tenant-registry/openapi.yaml
openapi: 3.0.0
info:
  title: Cloud Tenant Registry API
  version: 1.0.0
  description: Multi-tenant master data management service for FarmIQ
servers:
  - url: http://localhost:5121/api
    description: Local development server
paths:
...
  /v1/devices:
    get:
      summary: Get all devices for tenant
      tags: [Devices]
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: query
          schema:
            type: string
        - name: farmId
          in: query
          schema:
            type: string
        - name: barnId
          in: query
          schema:
            type: string
        - name: batchId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of devices
    post:
      summary: Create a new device
      tags: [Devices]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeviceRequest'
      responses:
        '201':
          description: Device created
        '409':
          description: Conflict (duplicate serialNumber)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateDeviceRequest:
      type: object
      required:
        - deviceType
      properties:
        deviceType:
          type: string
        serialNo:
          type: string
        farmId:
          type: string
          format: uuid
        barnId:
          type: string
          format: uuid
        batchId:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, inactive, maintenance]
        metadata:
          type: object
...